Split message feed code in two classes
* MessageFeedHandler keeps the message history, persists accross scenes and is the class through which messages are sent. It should be called server-side.
* UIMessageFeed receives messages from MessageFeedHandler and displays them on the client
