+++ Assets/BossRoom/Scenes/Startup.unity
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:304cbc5b539d79335e0412730810f144480c4c40b42c08a1066bae3067fd0aff
-size 33653
+oid sha256:ed2b475b5f8e8a6cfa3727face86a41bc8ae744d53f17ab4b77627e618951221
+size 34048
