+++ Assets/BossRoom/Scripts/Shared/Game/UI/MessageFeedHandler.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 2e044e5b3c1e0ad459826371abdb84f8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
